refactor: api주소 등을 상수화
