+++ src/components/ListItem/NotificationItem.tsx
@@ -1,5 +1,5 @@
 import { NOTIFICATION_STYLE as style } from '@/constants/styles';
-import type { NotificationItemProps } from '@/types/notification/notificationTypes';
+import type { NotificationItemProps } from '@/types/notificationTypes';
 
 export const NotificationItem = (props: NotificationItemProps) => {
   const { notificationId, content, createdAt } = props;
